PIM-4292: add item repository for mongo
