+++ src/Pim/Bundle/CatalogBundle/Resources/config/storage_driver/doctrine/mongodb-odm.yml
@@ -210,6 +210,7 @@ services:
 
     pim_catalog.repository.product_category:
         class: %pim_catalog.repository.product_category.class%
+        parent: pim_classifiation.repository.item_category
         arguments:
             - '@pim_catalog.object_manager.product'
             - %pim_catalog.entity.product.class%
